requirements.txt
Django
psycopg2-binary
python-dotenv
# pandas
# numpy